[FIX] update language stores from template
In Pootle 2.5 there was a update_against_templates mechanism, if a new
unit has been added to the default translation file, that updated all units
of the available languages.
Removed from Pootle 2.7 due to performance. As replacement, we now have
pot2po as local mechanism to update-merge the xliff files.

Resolves: #2